backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# AI/ML
torch==2.1.2
scikit-learn==1.4.0

# Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
